+++ widgets/mem_freebsd.go
@@ -57,5 +57,8 @@ func (self *MemWidget) updateSwapMemory() {
 			Used:        swapMemory.Used,
 			UsedPercent: swapMemory.UsedPercent,
 		})
+		if self.swapMetric != nil {
+			self.swapMetric.Set(swapMemory.UsedPercent)
+		}
 	}
 }